Moved structures from group data that could be re-used in scenes in a comon location: PersistentData.h
